database code less duplicate and more readable
