+++ databases/databaseClient.js
@@ -1,5 +1,5 @@
 var { Pool } = require("pg");
 
-exports.pool = new Pool({
+exports.client = new Pool({
   connectionString: process.env.CONNECTION_STRING
 });
