+++ databases/newsQuery.js
@@ -1,10 +1,9 @@
-var { pool } = require("./databaseURL"); 
-var { newsEventAndDateExist } = require("./newsEventAndDateExist");
-var { queryHadAnError } = require("./queryHadAnError");
+var { client } = require("./databaseURL"); 
+var { newsEventAndDateExist, queryHadAnError } = require("./helperFunctions");
 
 function getQuery(res){
   const sqlString = "SELECT * FROM Announcements";
-  pool.query(sqlString, function(err, result) {
+  client.query(sqlString, function(err, result) {
     if(queryHadAnError(err, res)) {return;}
     return res.status(200).send(result.rows);
   });
@@ -14,7 +13,7 @@ function postQuery(req, res){
   const {date, event} = req.query;
   if (!newsEventAndDateExist(date, event, res)) {return;}
   const sqlString = `INSERT INTO Announcements VALUES ('1', '${date}', '${event}')`;
-  pool.query(sqlString, function(err) {
+  client.query(sqlString, function(err) {
     if(queryHadAnError(err, res)) {return;}
     return res.status(201).send({message: `Event ${date}: ${event} created!`});
   });
@@ -24,7 +23,7 @@ function deleteQuery(req, res){
   const {date, event} = req.query;
   if (!newsEventAndDateExist(date, event, res)) {return;}
   const sqlString = `DELETE FROM Announcements WHERE date='${date}' AND event='${event}'`;
-  pool.query(sqlString, function(err) {
+  client.query(sqlString, function(err) {
     if(queryHadAnError(err, res)) {return;}
     return res.status(200).send({message: `Event ${date}: ${event} deleted!`});
   });
@@ -34,7 +33,7 @@ function putQuery(req, res){
   const {date, event, update} = req.query;
   if (!newsEventAndDateExist(date, event, res)) {return;}
   const sqlString = `UPDATE Announcements SET event='${event}' WHERE date='${date}' AND event='${event}'`;
-  pool.query(sqlString, function(err) {
+  client.query(sqlString, function(err) {
     if(queryHadAnError(err, res)) {return;}
     return res.status(200).send({message: `Event updated to ${date}: ${update}!`});
   });
